feat: Improve app splash screen
Doesn't hide the status bar, but makes it translucent instead. Displays a much better splash screen on devices with a top side notch, and on others, I think the status bar not hiding itself at all is better than being hidden just for a tiny amount of time at the start.