feat: Allow using tools:replace in Views' inject to not add e.g. android:id
This is essentially a copy of [1], where the low priority manifest is the one auto-generated by Sketchware Pro, and the high-priority one additions from the user done using the inject property.
The main idea is that this is a replacement for XML Command Blocks for minor changes like these.

Example: Add `tools:replace="android:textSize" android:textSize="@dimen/resource_text_size"` to a TextView's inject to use @dimen/resource_text_size as text size, not the value set in the View's text size property.

[1]: https://developer.android.com/studio/build/manage-manifests#attribute_markers